update immer remote react-redux, redux remove unused files rename all files to tsx
