+++ package.json
@@ -11,15 +11,13 @@
   },
   "dependencies": {
     "expo": "^36.0.0",
-    "immer": "^5.3.4",
+    "immer": "^5.3.6",
     "prop-types": "^15.7.2",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-36.0.0.tar.gz",
     "react-native-screens": "^2.0.0-beta.2",
     "react-native-web": "^0.12.1",
-    "react-redux": "^7.1.3",
-    "redux": "^4.0.5",
     "reselect": "^4.0.0"
   },
   "devDependencies": {
