--- src/data/index.js
@@ -1,9 +1,0 @@
-import game from './game';
-import frame from './frame';
-import pin from './pin';
-
-export default {
-  game,
-  frame,
-  pin
-};
