+++ src/data/index.tsx
@@ -1,0 +1,3 @@
+export { default as getGameData } from './game';
+export { default as getFrameData } from './frame';
+export { default as getPinData } from './pin';
