--- src/dataManagement/index.js
@@ -1,1 +1,0 @@
-export { default as game } from './game';
