--- src/reducers/reducers.tsx
@@ -1,8 +1,0 @@
-import { combineReducers } from 'redux';
-import game from './game';
-import app from './app';
-
-export default combineReducers({
-  app,
-  game
-});
