--- src/reducers/store.js
@@ -1,4 +1,0 @@
-import { createStore } from 'redux';
-import reducers from './reducers';
-
-export default createStore(reducers);
